Ai-Services/requirements.txt
fastapi
uvicorn
langchain
langchain-google-genai
langchain-community
chromadb
python-dotenv
pydantic
